+++ distributed/shuffle/_utils.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from typing import NewType
+
+ShuffleId = NewType("ShuffleId", str)
